VS2013 Compilation fixes
* Make Object pure virtual
* Include stdint for uint32_t type
* Include vector for vector type
* Use "int" instead of "unsigned short" since MSVC requires OpenMP index
variables be signed

+++ src/renderer.cpp
@@ -23,7 +23,7 @@ void Renderer::render(int samples) {
 
     // Main Loop
     #pragma omp parallel for schedule(dynamic, 1)       // OpenMP
-    for (unsigned short y=0; y<height; y++){
+    for (int y=0; y<height; y++){
         unsigned short Xi[3]={0,0,y*y*y};               // Stores seed for erand48
 
         fprintf(stderr, "\rRendering (%i samples): %.2f%% ",      // Prints
